:recycle: native query

+++ src/main/java/me/iqpizza6349/dote/domain/vote/ro/TeamRO.java
@@ -4,6 +4,8 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigInteger;
+
 @Getter
 @AllArgsConstructor
 public class TeamRO {
@@ -11,5 +13,5 @@ public class TeamRO {
     private final String name;
 
     @JsonProperty("vote_count")
-    private final int voteCount;
+    private final BigInteger voteCount;
 }
